Coloured noise, and unison on an oscillator
`Source::Noise` takes a `color` — white, pink or brown — and `Osc` takes
`voices` and `spread`. Both default to what the crate did before, and both
defaults are skipped on the way out, so no existing recipe's bytes or cached
bake move.

Pink is a bank of three staggered one-poles (Kellett's economical
coefficients), measured at -3.00 dB/octave with 1.1 dB of ripple across ten
octaves; a single one-pole was measured first and is not a colour at all. Brown
is a leaky integrator at 0.995, -5.89 dB/octave, bounded where an unleaky one
walks away. Each colour is scaled onto white's RMS.

Unison spreads `voices` copies evenly across `spread` cents, centred on the
written detune, each with its own start phase from the note seed, normalised by
the voice count. Capped at seven.

Refs #417

+++ crates/zimmer/src/core/karplus.rs
@@ -62,7 +62,7 @@ fn delay_len(freq: f32, rate: f32) -> usize {
 /// itself as `brightness` falls toward 0.
 fn excitation(len: usize, brightness: f32, seed: u64) -> Vec<f32> {
     let white: Vec<f32> = (0..len)
-        .map(|i| noise::white(i, EXCITATION_CHANNEL, seed))
+        .map(|i| noise::white(i as i64, EXCITATION_CHANNEL, seed))
         .collect();
     (0..len)
         .map(|i| {
